Make link to Code of Conduct absolute
The relative link to the Code of Conduct only works when viewing within
this repository. As the contents are pulled into other repos by Github
(e.g. https://github.com/alphagov/whitehall/security/policy) the
relative link leads to a non-existent page therefore giving a 404 error.
Making the link absolute so this is not a problem.